apps/cast/discord/pyproject.toml
[project]
name = "cast-discord"
version = "0.1.0"
description = "Discord bot for Cast integration"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.3",
    "aiohttp>=3.12.15",
    "python-dotenv>=1.1.1",
    "casting-discord-framework"
]

[tool.uv.sources]
casting-discord-framework = { workspace = true }

[project.scripts]
cast-discord = "casting.apps.cast.discord.run_bot:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0"
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "casting.apps.cast.discord"
namespace = true